fix: fixed layout bugs
